docker-build/requirements.txt
scikit-learn
pandas
matplotlib
seaborn
python-docx
openpyxl
pymupdf
